blogpost/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-taggit==3.0.0
Markdown==3.4.1
psycopg2-binary==2.9.3
sqlparse==0.4.3
tzdata==2022.4
